[Data] Rename `concurrency_solver.py` to `throughput_solver.py` (#61658)
Addresses review comment to rename this module to make it more
intuitive:
https://github.com/ray-project/ray/pull/61273#issuecomment-3974502162.
