+++ python/ray/data/tests/unit/test_throughput_solver.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ray.data._internal.cluster_autoscaler.concurrency_solver import (
+from ray.data._internal.cluster_autoscaler.throughput_solver import (
     allocate_resources,
     compute_optimal_throughput,
 )
